divide gigazine package
