+++ lib/sort.go
@@ -1,10 +1,12 @@
-package parse
+package lib
 
 import (
 	"sort"
+
+	"github.com/whywaita/gigani/parse"
 )
 
-func SortAnimeByTime(animes []Anime) []Anime {
+func SortAnimeByTime(animes []parse.Anime) []parse.Anime {
 	sort.Slice(animes, func(i, j int) bool {
 		return animes[i].StartDate.Before(animes[j].StartDate)
 	})
