+++ parse/gigazine/time.go
@@ -1,4 +1,4 @@
-package parse
+package gigazine
 
 import (
 	"errors"
