+++ parse/gigazine/time_test.go
@@ -1,4 +1,4 @@
-package parse
+package gigazine
 
 import (
 	"strings"
